+++ packages/generator-common/src/compiler.ts
@@ -240,6 +240,9 @@ export async function readCompilerOptions(
 
 function parseModuleResolutionKind(input: string): ModuleResolutionKind {
   const moduleResolution = input.toLowerCase();
+  if (moduleResolution === 'bundler') {
+    return ModuleResolutionKind.Bundler;
+  }
   if (moduleResolution === 'node') {
     return ModuleResolutionKind.Node10;
   }
@@ -264,7 +267,10 @@ function parseScriptTarget(input: string): ScriptTarget {
     es2019: ScriptTarget.ES2019,
     es2020: ScriptTarget.ES2020,
     es2021: ScriptTarget.ES2021,
-    es2022: ScriptTarget.ES2022
+    es2022: ScriptTarget.ES2022,
+    es2023: ScriptTarget.ES2023,
+    es2024: ScriptTarget.ES2024,
+    es2025: ScriptTarget.ES2025
   };
   if (mapping[input.toLowerCase()]) {
     return mapping[input.toLowerCase()];
@@ -279,11 +285,17 @@ function parseModuleKind(input: string): ModuleKind {
   const mapping: Record<string, ModuleKind> = {
     commonjs: ModuleKind.CommonJS,
     amd: ModuleKind.AMD,
+    umd: ModuleKind.UMD,
+    system: ModuleKind.System,
     es2015: ModuleKind.ES2015,
     es2020: ModuleKind.ES2020,
+    es2022: ModuleKind.ES2022,
     esnext: ModuleKind.ESNext,
     node16: ModuleKind.Node16,
-    nodenext: ModuleKind.NodeNext
+    node18: ModuleKind.Node18,
+    node20: ModuleKind.Node20,
+    nodenext: ModuleKind.NodeNext,
+    preserve: ModuleKind.Preserve
   };
 
   if (mapping[input.toLowerCase()]) {