+++ packages/openapi-generator/src/parser/document.spec.ts
@@ -2,7 +2,7 @@ import { createLogger } from '@sap-cloud-sdk/util';
 import { emptyDocument } from '../../test/test-util';
 import { parseOpenApiDocument } from './document';
 import * as api from './api';
-import type { ServiceOptions } from '@sap-cloud-sdk/generator-common/dist/options-per-service';
+import type { ServiceOptions } from '@sap-cloud-sdk/generator-common/internal';
 import type { OpenAPIV3 } from 'openapi-types';
 
 const options = { strictNaming: true, schemaPrefix: '', resolveExternal: true };